CrateAnalysis/requirements.txt
matplotlib==3.3.2
pandas_profiling==2.9.0
pandasgui==0.2.5.1
feather_format==0.4.1
pandas==1.1.2
scipy==1.5.2
numpy==1.18.5
